@@ -19,11 +19,11 @@
         }
     ],
     "require": {
-        "php" : "^5.5|^7.0",
-        "symfony/process": "^2.7|^3.0"
+        "php" : "^7.0",
+        "symfony/process": "^3.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8"
+        "phpunit/phpunit": "^5.5"
     },
     "autoload": {
         "psr-4": {
