@@ -14,6 +14,6 @@ class CannotSetParameter extends Exception
      */
     public static function conflictingParameters($name, $conflictName)
     {
-        return new static("Cannot set {$name} because it conflicts with parameter {$conflictName}.");
+        return new static("Cannot set `{$name}` because it conflicts with parameter `{$conflictName}`.");
     }
 }
