@@ -13,6 +13,6 @@ class CannotStartDump extends Exception
      */
     public static function emptyParameter($name)
     {
-        return new static("Parameter {$name} cannot be empty.");
+        return new static("Parameter `{$name}` cannot be empty.");
     }
 }
